moved schema out of the Facade
