+++ genus-library/src/main/scala/co/topl/genusLibrary/GenusGraphMetadata.scala
@@ -1,0 +1,89 @@
+package co.topl.genusLibrary
+
+import com.orientechnologies.orient.core.metadata.schema.OType
+import com.tinkerpop.blueprints.impls.orient.{OrientEdgeType, OrientGraphNoTx, OrientVertexType}
+
+/**
+ * Metadata describing the schema used for the Genus graph in OrientDB
+ */
+class GenusGraphMetadata (val session: OrientGraphNoTx) {
+  import GenusGraphMetadata._
+
+  val addressVertexType: OrientVertexType = session.createVertexType("Address")
+  configureAddressVertexType()
+
+  val boxStateVertexType: OrientVertexType = session.createVertexType("BoxState")
+  // box states have no properties to configure
+
+  val blockHeaderVertexType: OrientVertexType = session.createVertexType("BlockHeader")
+  configureBlockHeaderVertexType()
+
+  val blockBodyVertexType: OrientVertexType = session.createVertexType("BlockBody")
+  val boxVertexType: OrientVertexType = session.createVertexType("Box")
+  val transactionVertexType: OrientVertexType = session.createVertexType("Transaction")
+
+  val currentBoxStateEdgeType: OrientEdgeType = session.createEdgeType("CurrentBoxState")
+  val prevToNextBoxStateEdgeType: OrientEdgeType = session.createEdgeType("PrevToNext")
+  val stateToBoxEdgeType: OrientEdgeType = session.createEdgeType("StateToBox")
+  val inputEdgeType: OrientEdgeType = session.createEdgeType("Input")
+  val outputEdgeType: OrientEdgeType = session.createEdgeType("Output")
+
+  private def configureAddressVertexType(): Unit = {
+    addressVertexType
+      .createProperty("base58Address", OType.STRING)
+      .setMandatory(true)
+      .setReadonly(true)
+      .setNotNull(true)
+      .setRegexp(TypedEvidenceRegex)
+    session.createIndex(
+      "base58Address",
+      classOf[Vertex],
+      new Parameter("type", "UNIQUE"),
+      new Parameter("class", addressVertexType.getName)
+    )
+  }
+
+  def configureBlockHeaderVertexType(): Unit = {
+    blockHeaderVertexType
+      .createProperty("blockId", OType.INTEGER)
+      .setMandatory(true)
+      .setReadonly(true)
+      .setNotNull(true)
+    session.createIndex(
+      "blockId",
+      classOf[Vertex],
+      new Parameter("type", "UNIQUE"),
+      new Parameter("class", blockHeaderVertexType.getName)
+    )
+
+    blockHeaderVertexType
+      .createProperty("parentHeaderId", OType.INTEGER)
+      .setMandatory(false)
+      .setReadonly(true)
+      .setNotNull(true)
+
+    blockHeaderVertexType
+      .createProperty("parentSlot", OType.LONG)
+      .setMandatory(false)
+      .setReadonly(true)
+      .setNotNull(true)
+
+    blockHeaderVertexType
+      .createProperty("txRoot", OType.BINARY) // 32 Bytes
+      .setMandatory(false)
+      .setReadonly(true)
+      .setNotNull(false)
+
+    blockHeaderVertexType
+      .createProperty("bloomFilter", OType.BINARY) // 256 Bytes
+      .setMandatory(false)
+      .setReadonly(true)
+      .setNotNull(false)
+
+  }
+}
+
+object GenusGraphMetadata {
+  private val TypedEvidenceRegex = "^[123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz]{33,46}$"
+
+}
