+++ genus-library/src/main/scala/co/topl/genusLibrary/OrientDBFacade.scala
@@ -1,10 +1,8 @@
 package co.topl.genusLibrary
 
 import co.topl.typeclasses.Log._
-import com.orientechnologies.orient.core.metadata.schema.OType
 import com.orientechnologies.orient.core.sql.OCommandSQL
-import com.tinkerpop.blueprints.impls.orient.{OrientEdgeType, OrientGraphFactory, OrientVertexType}
-import com.tinkerpop.blueprints.{Parameter, Vertex}
+import com.tinkerpop.blueprints.impls.orient.{OrientGraphFactory, OrientGraphNoTx}
 import com.typesafe.scalalogging.Logger
 
 import java.io.{File, FileInputStream, FileOutputStream}
@@ -25,90 +23,15 @@ class OrientDBFacade(dir: File, password: String) {
   private val factory = new OrientGraphFactory(fileToPlocalUrl(dir), dbUserName, password)
 
   @unused
-  private val schemaMetadata = initializeDatabase(factory, password)
+  private val graphMetadata = initializeDatabase(factory, password)
 
   private def initializeDatabase(factory: OrientGraphFactory, password: String) = {
-    val session = factory.getNoTx
+    val session: OrientGraphNoTx = factory.getNoTx
     try {
       logger.info("Changing password")
       session.command(new OCommandSQL(s"UPDATE OUser SET password='$password' WHERE name='$dbUserName'")).execute()
       logger.info("Configuring Schema")
-      new {
-        val addressVertexType: OrientVertexType = session.createVertexType("Address")
-        configureAddressVertexType()
-
-        val boxStateVertexType: OrientVertexType = session.createVertexType("BoxState")
-        // box states have no properties to configure
-
-        val blockHeaderVertexType: OrientVertexType = session.createVertexType("BlockHeader")
-        configureBlockHeaderVertexType()
-
-        val blockBodyVertexType: OrientVertexType = session.createVertexType("BlockBody")
-        val boxVertexType: OrientVertexType = session.createVertexType("Box")
-        val transactionVertexType: OrientVertexType = session.createVertexType("Transaction")
-
-        val currentBoxStateEdgeType: OrientEdgeType = session.createEdgeType("CurrentBoxState")
-        val prevToNextBoxStateEdgeType: OrientEdgeType = session.createEdgeType("PrevToNext")
-        val stateToBoxEdgeType: OrientEdgeType = session.createEdgeType("StateToBox")
-        val inputEdgeType: OrientEdgeType = session.createEdgeType("Input")
-        val outputEdgeType: OrientEdgeType = session.createEdgeType("Output")
-
-        private def configureAddressVertexType(): Unit = {
-          addressVertexType
-            .createProperty("base58Address", OType.STRING)
-            .setMandatory(true)
-            .setReadonly(true)
-            .setNotNull(true)
-            .setRegexp(TypedEvidenceRegex)
-          session.createIndex(
-            "base58Address",
-            classOf[Vertex],
-            new Parameter("type", "UNIQUE"),
-            new Parameter("class", addressVertexType.getName)
-          )
-        }
-
-        def configureBlockHeaderVertexType(): Unit = {
-          blockHeaderVertexType
-            .createProperty("blockId", OType.INTEGER)
-            .setMandatory(true)
-            .setReadonly(true)
-            .setNotNull(true)
-          session.createIndex(
-            "blockId",
-            classOf[Vertex],
-            new Parameter("type", "UNIQUE"),
-            new Parameter("class", blockHeaderVertexType.getName)
-          )
-
-          blockHeaderVertexType
-            .createProperty("parentHeaderId", OType.INTEGER)
-            .setMandatory(false)
-            .setReadonly(true)
-            .setNotNull(true)
-
-          blockHeaderVertexType
-            .createProperty("parentSlot", OType.LONG)
-            .setMandatory(false)
-            .setReadonly(true)
-            .setNotNull(true)
-
-          blockHeaderVertexType
-            .createProperty("txRoot", OType.BINARY) // 32 Bytes
-            .setMandatory(false)
-            .setReadonly(true)
-            .setNotNull(false)
-
-          blockHeaderVertexType
-            .createProperty("bloomFilter", OType.BINARY) // 256 Bytes
-            .setMandatory(false)
-            .setReadonly(true)
-            .setNotNull(false)
-
-          blockHeaderVertexType
-        }
-
-      }
+      new GenusGraphMetadata(session)
     } finally
       session.shutdown()
   }
@@ -132,8 +55,6 @@ object OrientDBFacade {
 
   private val charsetUtf8: Charset = Charset.forName("UTF-8")
   private val passwdLength = 30
-
-  private val TypedEvidenceRegex = "^[123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz]{33,46}$"
 
   /**
    * Create an instance of OrientDBFacade
